# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: DocFlow API
  version: 1.0.0
  description: |
    DocFlow 文档管理服务 API。

    ## 认证
    - 业务端点要求 `Authorization: Bearer <access_token>`（HS256 JWT，subject 为用户 UUID）。
    - 登录/刷新通过 HttpOnly Cookie `refresh_token`（Path=/api/v1/auth/refresh，SameSite=Lax）轮换会话。

    ## 错误与限流
    - 错误响应统一为 `{error, code?}`；`code` 为机器可读错误码（如 NAME_CONFLICT、QUARANTINED）。
    - 认证接口按 user_id+IP 每分钟限流；登录按 IP+邮箱前缀哈希更严格限流；
      公开分享接口按 IP 独立限流；超限返回 429 并带 Retry-After。
    - 文件元数据响应带 ETag（基于 updated_at）。
servers:
  - url: /api/v1
    description: API 基础路径（契约测试据此归一化路由）
security:
  - bearerAuth: []
...
paths:
...
  /files/batch/move:
    post:
      tags: [Batch]
      operationId: batchMove
      summary: 批量移动（部分成功）
      description: >-
        把多个文件/目录移动到 target_parent_id（缺省为个人根目录）下。
        目标目录整体校验失败（不存在 404 / 无写权限 403）时整个请求失败；
        其余逐项执行、逐项收集结果 [{id,ok,error_code}]，单项失败不回滚
        其他项（部分成功语义），目标目录同名冲突项跳过（NAME_CONFLICT）。
        移动继承目标目录作用域（跨空间移动更新 space_id 并校验目标空间
        配额）。根目录与
        移入自身/后代（INVALID_TARGET）拒绝；移动后子树最深深度超过
        folder.max_depth（默认 32）时逐项 DEPTH_LIMIT 跳过。file_ids 去重、
        上限经 batch.max_items 热读取（默认 100）。
        可选 Idempotency-Key：同 key+同请求体 60s 内重放缓存响应
        （Idempotency-Replayed: true），同 key 不同请求体 409；
        缓存为进程内实现，多实例部署下不跨实例共享。
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
          description: 幂等键（batch 三端点通用语义）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                target_parent_id:
                  { type: string, format: uuid, description: 目标目录 ID；缺省为个人根目录 }
      responses:
        '200':
          description: 逐项结果（含失败项 error_code）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchResults' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Idempotency-Key 已被不同请求体使用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /files/batch/download:
    post:
      tags: [Batch]
      operationId: batchDownload
      summary: 批量安全 ZIP 下载
      description: 最多 100 个可读文件，服务端仅打包 available blob，文件名加 UUID 前缀防冲突与路径穿越。
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 100
      responses:
        '200':
          description: ZIP 流
          content:
            application/zip:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /files/batch/trash:
    post:
      tags: [Batch]
      operationId: batchTrash
      summary: 批量软删除（部分成功）
      description: >-
        批量移入回收站（软删除）；根目录跳过（ROOT），个人文件仅 owner、
        团队文件要求成员写权限（FORBIDDEN）。逐项结果语义与上限同 move。
        Idempotency-Key 语义同 move。
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 100
      responses:
        '200':
          description: 逐项结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchResults' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Idempotency-Key 已被不同请求体使用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /files/batch/restore:
    post:
      tags: [Batch]
      operationId: batchRestore
      summary: 批量恢复（部分成功）
      description: >-
        批量从回收站恢复，逐项复用单文件 Restore 语义：原父目录已删除
        （PARENT_DELETED）/ 名称冲突（NAME_CONFLICT）/ 未删除（NOT_DELETED）
        记入 per-item 结果，不静默改名。Idempotency-Key 语义同 move。
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 100
      responses:
        '200':
          description: 逐项结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchResults' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Idempotency-Key 已被不同请求体使用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Authorization: Bearer <token>，两类凭证等价可用：
        ① HS256 JWT（subject 为用户 UUID，过期时间以服务配置为准）；
        ② PAT 个人访问令牌：dfpat_ + 43 字符 URL-safe base64（经
        POST /tokens 创建，撤销/过期即失效；无 refresh 语义）。
  responses:
    BadRequest:
      description: 请求参数或请求体非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 缺失或无效的访问令牌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 无权限执行该操作（含只读目录、根目录保护、对象不可用等）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在或不属于当前用户（不泄露存在性）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: 触发限流；响应头 Retry-After 为当前窗口剩余秒数
      headers:
        Retry-After:
          description: 建议等待秒数
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 人类可读错误信息
        code:
          type: string
          description: >-
            机器可读错误码；当前实现使用 NAME_CONFLICT / NOT_DELETED / PARENT_DELETED /
            QUARANTINED / FAILED / TOTP_REQUIRED（POST /auth/login 对已启用两步
            验证账号的 401，error=totp_required）/ ACCOUNT_LOCKED（登录与
            refresh 遇锁定期账号的 423）/ QUOTA_EXCEEDED（上传建会话配额
            超限的 403）/ CSRF_REJECTED（refresh/logout 同源校验失败的 403）
        status:
          type: string
          description: 上传相关错误附带的会话状态（仅 uploads complete 端点）
...
    BatchResultItem:
      type: object
      description: 批量操作单项结果；ok=false 时附机器可读错误码
      properties:
        id: { type: string, format: uuid }
        ok: { type: boolean }
        error_code:
          type: string
          description: >-
            NOT_FOUND / FORBIDDEN / ROOT / NAME_CONFLICT / INVALID_TARGET /
            PARENT_DELETED / NOT_DELETED / INTERNAL
    BatchResults:
      type: object
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/BatchResultItem' }
...